♻️ refactor: landing page vite refactorying
