+++ src/@shared/Footer.css.ts
@@ -1,0 +1,8 @@
+import { style } from "@vanilla-extract/css";
+
+export const footer = style({
+  padding: "0.5rem 0",
+  fontSize: "1.1rem",
+  fontWeight: "300",
+  textAlign: "center",
+});
