+++ src/@shared/Footer.tsx
@@ -1,0 +1,9 @@
+import * as css from "./Footer.css";
+
+export function Footer() {
+  return (
+    <footer className={css.footer}>
+      Every deployment from the 2023 edition of Jaam Toast.
+    </footer>
+  );
+}
