+++ src/@shared/Header.css.ts
@@ -1,5 +1,5 @@
 import { style } from "@vanilla-extract/css";
-import { s } from "../../styles";
+import { s } from "../styles";
 import { gmarketSans } from "src/styles/font.css";
 
 export const container = style([
